Move pw tests into multiple files
