+++ tests/Playwright/tests/free-trial-subscriptions.spec.js
@@ -1,10 +1,15 @@
 const { test, expect } = require( '@playwright/test' );
 const { loginAsCustomer } = require( './utils/user' );
 const { openPaypalPopup, loginIntoPaypal } = require( './utils/paypal-popup' );
+const { serverExec } = require( './utils/server' );
 
-test( 'PayPal logged-in user free trial subscription without payment token', async ( {
+test( 'PayPal logged-in user free trial subscription without payment token with shipping callback enabled', async ( {
 	page,
 } ) => {
+	await serverExec(
+		'wp pcp settings update blocks_final_review_enabled false'
+	);
+
 	await loginAsCustomer( page );
 
 	await page.goto( '/product/free-trial' );
@@ -14,8 +19,6 @@ test( 'PayPal logged-in user free trial subscription without payment token', asy
 	const popup = await openPaypalPopup( page );
 	await loginIntoPaypal( popup );
 	popup.locator( '#consentButton' ).click();
-
-	await page.click( 'text=Proceed to PayPal' );
 
 	const title = await page.locator( '.entry-title' );
 	await expect( title ).toHaveText( 'Order received' );
